requirements.txt
pandas>=2.3.2
streamlit>=1.49.1
plotly>=5.0.0
openai>=1.0.0
python-dotenv>=1.0.0
transformers>=4.56.0
sentence-transformers>=5.1.0
scikit-learn>=1.7.1
matplotlib>=3.10.6
seaborn>=0.13.2
fuzzywuzzy>=0.18.0
numpy>=2.3.2
jupyter>=1.0.0
jupyterlab>=4.4.6
notebook>=7.4.5
python-levenshtein>=0.12.0
requests>=2.32.5
